pyproject.toml
[tool.poetry]
name = "stock-market-data"
version = "0.1.0"
description = ""
authors = ["Shabir Hussain <shabir.hussain326@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.0"
djangorestframework = "^3.15.1"
drf-yasg = "^1.21.7"
yfinance = "^0.2.38"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120